Fixes #37891 - Add enable-puppet8 parameter to provisioning templates
Previously the enable-official-puppet8-repo parameter was added, but for
users who supply their own packages (like Katello users) and still want
to use AIO packaging need a parameter. Technically they could use
enable-puppet5, enable-puppet6 or enable-puppet7 but this is
inconsistent. It would be better to add a new version independent
parameter, but this first aims for consistency.

Fixes: c35092bdba5d ("fixes #36939 - Allow for deployment of puppet 8")